Vitest workspace (#12042)
* workspace

* add test:dist

* apps TODO

* debug

* test

* debug

* full logs

* .

* remove script containing 'with', rollup cannot handle it

* glob

* cleanup

* root coverage

* test

* attempt again

* wip

* rework

* ???

* revert

* lockfile

* yarn dedupe

* exclude

* ??

* test

* mts

* flatten config

* revert yarn.lock

* update yarn.lock

* log

* more debugging

* test

* revert reference to code vite can't process

* add vitest.config.ts to .dockerignore

* revert debugging changes

* simplify config

* cleanup

* bump

* fixup

* disable isolate

* missing coverage

* more explicit test target

* Attempt without specific includes

* bump vitest

* yarn upgrade-interactive for vitest

* Share config between workspace root and apps/prairielearn vitest

* Remove redundant isolate config

* filter by project
