+++ apps/grader-host/package.json
@@ -43,11 +43,11 @@
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^22.15.29",
     "@types/tmp": "^0.2.6",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "nodemon": "^3.1.10",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
     "typescript-cp": "^0.1.9",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
