+++ apps/prairielearn/package.json
@@ -250,7 +250,7 @@
     "@types/tmp": "^0.2.6",
     "@types/unzipper": "^0.10.11",
     "@types/uuid": "^10.0.0",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "axe-core": "^4.10.3",
     "get-port": "^7.1.0",
     "jsdom": "^26.1.0",
@@ -258,7 +258,7 @@
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
     "typescript-cp": "^0.1.9",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   },
   "nodemonConfig": {
     "ignore": [
