+++ apps/workspace-host/package.json
@@ -45,13 +45,13 @@
   },
   "devDependencies": {
     "@types/node": "^22.15.29",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "fast-glob": "^3.3.3",
     "nodemon": "^3.1.10",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
     "typescript-cp": "^0.1.9",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   },
   "nodemonConfig": {
     "ext": "js,jsx,ts,tsx,json,sql"
