+++ package.json
@@ -22,6 +22,7 @@
     "release": "turbo run build && node scripts/fix-workspace-versions-before-publish.mjs && changeset publish --no-git-tag",
     "start": "yarn workspace @prairielearn/prairielearn start",
     "start-workspace-host": "yarn workspace @prairielearn/workspace-host start",
+    "test": "vitest run --coverage",
     "version": "changeset version && YARN_ENABLE_IMMUTABLE_INSTALLS=false yarn"
   },
   "dependencies": {
@@ -38,6 +39,7 @@
     "@changesets/cli": "^2.29.4",
     "@eslint-react/eslint-plugin": "^1.51.0",
     "@prairielearn/prettier-plugin-sql": "workspace:^",
+    "@vitest/coverage-v8": "^3.2.2",
     "@vitest/eslint-plugin": "^1.2.1",
     "dependency-cruiser": "^16.10.2",
     "eslint": "^9.28.0",
@@ -56,7 +58,8 @@
     "s3rver": "^3.7.1",
     "turbo": "^2.5.4",
     "typescript": "^5.8.3",
-    "typescript-eslint": "^8.33.1"
+    "typescript-eslint": "^8.33.1",
+    "vitest": "^3.2.2"
   },
   "resolutions": {
     "notebookjs/jsdom": "^25.0.0",
