+++ packages/config/package.json
@@ -21,10 +21,10 @@
     "@types/fs-extra": "^11.0.4",
     "@types/lodash": "^4.17.17",
     "@types/node": "^22.15.29",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "tmp-promise": "^3.0.3",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
