+++ packages/error/package.json
@@ -20,9 +20,9 @@
     "@prairielearn/html": "workspace:^",
     "@prairielearn/tsconfig": "workspace:^",
     "@types/node": "^22.15.29",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
