+++ packages/eslint-plugin-prairielearn/package.json
@@ -19,9 +19,9 @@
     "@prairielearn/tsconfig": "workspace:^",
     "@types/node": "^22.15.29",
     "@typescript-eslint/rule-tester": "^8.33.1",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
