+++ packages/html-ejs/package.json
@@ -21,9 +21,9 @@
     "@prairielearn/tsconfig": "workspace:^",
     "@types/ejs": "^3.1.5",
     "@types/node": "^22.15.29",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
