+++ packages/markdown/package.json
@@ -32,10 +32,10 @@
     "@types/mdast": "^4.0.4",
     "@types/node": "^22.15.29",
     "@types/unist": "^3.0.3",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "mdast-util-math": "^3.0.0",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
