+++ packages/postgres/package.json
@@ -28,9 +28,9 @@
     "@prairielearn/tsconfig": "workspace:^",
     "@types/multipipe": "^3.0.5",
     "@types/node": "^22.15.29",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
