+++ packages/session/package.json
@@ -31,7 +31,7 @@
     "@types/on-headers": "^1.0.3",
     "@types/set-cookie-parser": "^2.4.10",
     "@types/uid-safe": "^2.1.5",
-    "@vitest/coverage-v8": "^3.2.1",
+    "@vitest/coverage-v8": "^3.2.2",
     "express": "^4.21.2",
     "fetch-cookie": "^3.1.0",
     "node-fetch": "^3.3.2",
@@ -39,6 +39,6 @@
     "source-map-support": "^0.5.21",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
-    "vitest": "^3.2.1"
+    "vitest": "^3.2.2"
   }
 }
